Simpler code structure
